% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercion.R, R/ShapeMap.R, R/AxialShapeGraph.R,
%   R/SegmentShapeGraph.R
\name{alcyon-coercion}
\alias{alcyon-coercion}
\alias{coerce,sf,ShapeMap-method}
\alias{coerce,ShapeMap,sf-method}
\alias{coerce,ShapeMap,AxialShapeGraph-method}
\alias{coerce,sf,AxialShapeGraph-method}
\alias{coerce,ShapeMap,SegmentShapeGraph-method}
\alias{coerce,sf,SegmentShapeGraph-method}
\title{Coercion between sf and alcyon map classes}
\usage{
\S4method{coerce}{sf,ShapeMap}(from, to = "ShapeMap", strict = TRUE)

\S4method{coerce}{ShapeMap,sf}(from, to = "sf", strict = TRUE)

\S4method{coerce}{ShapeMap,AxialShapeGraph}(from, to = "AxialShapeGraph", strict = TRUE)

\S4method{coerce}{sf,AxialShapeGraph}(from, to = "AxialShapeGraph", strict = TRUE)

\S4method{coerce}{ShapeMap,SegmentShapeGraph}(from, to = "SegmentShapeGraph", strict = TRUE)

\S4method{coerce}{sf,SegmentShapeGraph}(from, to = "SegmentShapeGraph", strict = TRUE)
}
\arguments{
\item{from}{the object to coerce.}

\item{to}{the target class. Supplied by \code{\link[methods]{as}} and not
used by the method bodies.}

\item{strict}{logical, part of the \code{coerce} generic signature. Not used
by the method bodies.}
}
\description{
Direct conversions between \code{sf} objects and the map classes provided by
this package. For ShapeMap -> Axial -> Segment see
\link{axialToSegmentShapeGraph}.

This is a direct conversion, for ShapeMap -> Axial -> Segment see
\link{axialToSegmentShapeGraph}

This is a direct conversion, for ShapeMap -> Axial -> Segment see
\link{axialToSegmentShapeGraph}
}
\seealso{
Other ShapeMap:
\code{\link{ShapeMap-class}}

Other AxialShapeGraph:
\code{\link{AxialShapeGraph-class}}

Other SegmentShapeGraph:
\code{\link{SegmentShapeGraph-class}}
}
\concept{AxialShapeGraph}
\concept{SegmentShapeGraph}
\concept{ShapeMap}
