% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllLineShapeGraph.R
\docType{class}
\name{AllLineShapeGraph-class}
\alias{AllLineShapeGraph-class}
\title{All-line Axial ShapeGraph}
\description{
A representation of sala's All-line ShapeGraph in R. Holds onto a sala
All-line ShapeGraph pointer and operates on that
}
