% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxialShapeGraph.R
\name{AxialShapeGraph_subset}
\alias{AxialShapeGraph_subset}
\alias{[.AxialShapeGraph}
\alias{[<-.AxialShapeGraph}
\title{Subset AxialShapeGraph objects}
\usage{
\method{[}{AxialShapeGraph}(x, ...)

\method{[}{AxialShapeGraph}(x, ...) <- value
}
\arguments{
\item{x}{object of class \code{AxialShapeGraph} passed to \code{stars[]}}

\item{...}{other parameters passed to \code{stars[] <- }}

\item{value}{value to be passed to \code{sf[] <- }}
}
\description{
Subsetting AxialShapeGraph objects essentially passes the data to sf.
See \link[sf]{sf}
}
