% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfConversions.R
\name{axialToSegmentShapeGraph}
\alias{axialToSegmentShapeGraph}
\title{Axial to Segment ShapeGraph}
\usage{
axialToSegmentShapeGraph(axialShapeGraph, stubRemoval = NULL)
}
\arguments{
\item{axialShapeGraph}{An Axial ShapeGraph}

\item{stubRemoval}{Rremove stubs of axial lines shorter than this
percentage (for example provide 0.4 for 40\%)}
}
\value{
A new Segment ShapeGraph
}
\description{
Convert an Axial ShapeGraph to a Segment ShapeGraph
}
