% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVGA.R
\name{blockLines}
\alias{blockLines}
\title{Block lines on a PointMap}
\usage{
blockLines(pointMap, lineStringMap, verbose = FALSE)
}
\arguments{
\item{pointMap}{The input PointMap}

\item{lineStringMap}{Map of lines, either a ShapeMap, or an sf lineString map}

\item{verbose}{Optional. Show more information of the process.}
}
\description{
Takes a PointMap and a ShapeMap with lines and blocks the cells on the
PointMap where the lines pass.
}
