% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\docType{methods}
\name{connections,PointMap-method}
\alias{connections,PointMap-method}
\title{Get the PointMap connections}
\usage{
\S4method{connections}{PointMap}(map)
}
\arguments{
\item{map}{A PointMap}
}
\value{
A matrix with the connected refs
}
\description{
Get the PointMap connections
}
