% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isovist.R
\name{isovist2pts}
\alias{isovist2pts}
\title{Create isovists using two points}
\usage{
isovist2pts(boundaryMap, x, y, toX, toY, viewAngle, verbose = FALSE)
}
\arguments{
\item{boundaryMap}{A ShapeMap with lines designating the isovist boundaries}

\item{x}{X coordinate of the origin points}

\item{y}{Y coordinate of the origin points}

\item{toX}{X coordinate of the target points}

\item{toY}{Y coordinate of the taget points}

\item{viewAngle}{The angle signifying the isovist's field of view}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
A ShapeMap with the isovist polygons
}
\description{
Create one or more isovists at particular points, given another point for
direction and an angle for field of view
}
