% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeMap.R
\name{LatticeMap-class}
\alias{LatticeMap-class}
\alias{LatticeMap}
\title{LatticeMap}
\description{
A representation of sala's LatticeMap in R. Holds onto a sala LatticeMap pointer
and operates on that
}
\concept{LatticeMap}
