% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeMap.R
\name{LatticeMap_subset}
\alias{LatticeMap_subset}
\alias{[.LatticeMap}
\alias{[<-.LatticeMap}
\title{Subset LatticeMap objects}
\usage{
\method{[}{LatticeMap}(x, ...)

\method{[}{LatticeMap}(x, ...) <- value
}
\arguments{
\item{x}{object of class \code{LatticeMap} passed to \code{stars[]}}

\item{...}{other parameters passed to \code{stars[] <- }}

\item{value}{value to be passed to \code{stars[] <- }}
}
\description{
Subsetting LatticeMap objects essentially passes the data to stars
See \link[stars]{stars_subset}
}
