% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxialShapeGraph.R
\docType{methods}
\name{linkCoords,AxialShapeGraph-method}
\alias{linkCoords,AxialShapeGraph-method}
\title{Link two Axial Lines (coordinates)}
\usage{
\S4method{linkCoords}{AxialShapeGraph}(map, fromX, fromY, toX, toY)
}
\arguments{
\item{map}{An Axial ShapeGraph}

\item{fromX}{X coordinate of the first link point}

\item{fromY}{Y coordinate of the first link point}

\item{toX}{X coordinate of the second link point}

\item{toY}{Y coordinate of the second link point}
}
\description{
Link two locations on an Axial ShapeGraph using the point coordinates
}
