% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{linkCoords}
\alias{linkCoords}
\title{Link map points/lines as if selecting them using points}
\usage{
linkCoords(map, fromX, fromY, toX, toY, copyMap = TRUE)
}
\arguments{
\item{map}{A sala map}

\item{fromX}{X coordinate of the origin point}

\item{fromY}{Y coordinate of the origin point}

\item{toX}{X coordinate of the target point}

\item{toY}{Y coordinate of the target point}

\item{copyMap}{Optional. Copy the internal sala map}
}
\value{
A new map with linked points/lines
}
\description{
Link map points/lines as if selecting them using points
}
