% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\docType{methods}
\name{linkRefs,PointMap-method}
\alias{linkRefs,PointMap-method}
\title{Link two Axial Lines (refs)}
\usage{
\S4method{linkRefs}{PointMap}(map, fromRef, toRef)
}
\arguments{
\item{map}{A PointMap}

\item{fromRef}{Ref of the first link line}

\item{toRef}{Ref of the second link line}
}
\description{
Link two lines on an Axial ShapeGraph using their refs
}
