% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxialShapeGraph.R
\docType{methods}
\name{links,AxialShapeGraph-method}
\alias{links,AxialShapeGraph-method}
\title{Get the Axial ShapeGraph links}
\usage{
\S4method{links}{AxialShapeGraph}(map)
}
\arguments{
\item{map}{An Axial ShapeGraph}
}
\value{
A matrix with the linked refs
}
\description{
Get the Axial ShapeGraph links
}
