% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\docType{methods}
\name{links,PointMap-method}
\alias{links,PointMap-method}
\title{Get the PointMap links}
\usage{
\S4method{links}{PointMap}(map)
}
\arguments{
\item{map}{A PointMap}
}
\value{
A matrix with the linked refs
}
\description{
Get the PointMap links
}
