% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFewestLineMap.R
\name{makeAllLineMap}
\alias{makeAllLineMap}
\title{Create an All-line Map}
\usage{
makeAllLineMap(boundsMap, seedX, seedY, verbose = FALSE)
}
\arguments{
\item{boundsMap}{The boundary ShapeMap to create the all-line map in}

\item{seedX}{X coordinate of the seed (the point that initiates the process)}

\item{seedY}{Y coordinate of the seed (the point that initiates the process)}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
An All-line Axial ShapeGraph
}
\description{
Create an All-line Map
}
