% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{makeColour}
\alias{makeColour}
\alias{makeDepthmapClassicColour}
\alias{makeAxmanesqueColour}
\alias{makePurpleOrangeColour}
\alias{makeBlueRedColour}
\alias{makeGreyScaleColour}
\alias{makeNiceHSBColour}
\title{Single Colour from depthmapX's Palettes}
\usage{
makeDepthmapClassicColour(value, rangeMin = 0, rangeMax = 1)

makeAxmanesqueColour(value, rangeMin = 0, rangeMax = 1)

makePurpleOrangeColour(value, rangeMin = 0, rangeMax = 1)

makeBlueRedColour(value, rangeMin = 0, rangeMax = 1)

makeGreyScaleColour(value, rangeMin = 0, rangeMax = 1)

makeNiceHSBColour(value, rangeMin = 0, rangeMax = 1)
}
\arguments{
\item{value}{Value within the min/max range to take}

\item{rangeMin}{The min value of the range}

\item{rangeMax}{The max value of the range}
}
\value{
Returns a single colour.
}
\description{
Create a single colour from depthmapX's palettes.
}
\examples{
makeDepthmapClassicColour(0.2, 0, 1)
makeAxmanesqueColour(0.2, 0, 1)
makePurpleOrangeColour(0.2, 0, 1)
makeBlueRedColour(0.2, 0, 1)
makeGreyScaleColour(0.2, 0, 1)
makeNiceHSBColour(0.2, 0, 1)
}
