% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVGA.R
\name{makeVGAGraph}
\alias{makeVGAGraph}
\title{Create a graph between visible cells in the PointMap}
\usage{
makeVGAGraph(
  pointMap,
  boundaryGraph = FALSE,
  maxVisibility = NA,
  verbose = FALSE
)
}
\arguments{
\item{pointMap}{The input PointMap}

\item{boundaryGraph}{Only create a graph on the boundary cells}

\item{maxVisibility}{Limit how far two cells can be to be connected}

\item{verbose}{Optional. Show more information of the process.}
}
\description{
Create a graph between visible cells in the PointMap
}
