% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeMap.R
\docType{methods}
\name{name,LatticeMap-method}
\alias{name,LatticeMap-method}
\title{Get the LatticeMap name}
\usage{
\S4method{name}{LatticeMap}(map)
}
\arguments{
\item{map}{A LatticeMap}
}
\value{
The name of the LatticeMap as a string
}
\description{
Get the LatticeMap name
}
\examples{
mifFile <- system.file(
    "extdata", "testdata", "gallery",
    "gallery_lines.mif",
    package = "alcyon"
  )
  sfMap <- st_read(mifFile,
    geometry_column = 1L, quiet = TRUE
  )
  latticeMap <- makeVGALatticeMap(
    sfMap,
    gridSize = 0.04,
    fillX = 3.01,
    fillY = 6.7,
    maxVisibility = NA,
    boundaryGraph = FALSE,
    verbose = FALSE
  )
name(latticeMap)
}
