% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\docType{methods}
\name{name,PointMap-method}
\alias{name,PointMap-method}
\title{Get the PointMap name}
\usage{
\S4method{name}{PointMap}(map)
}
\arguments{
\item{map}{A PointMap}
}
\value{
The name of the PointMap as a string
}
\description{
Get the PointMap name
}
