% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeMap.R
\name{name,ShapeMap-method}
\alias{name,ShapeMap-method}
\title{Get the ShapeMap name}
\usage{
\S4method{name}{ShapeMap}(map)
}
\arguments{
\item{map}{A ShapeMap}
}
\value{
The name of the ShapeMap as a string
}
\description{
Get the ShapeMap name
}
