% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneToAllTraverse.R
\name{oneToAllTraverse}
\alias{oneToAllTraverse}
\title{One-to-all traversal}
\usage{
oneToAllTraverse(
  map,
  traversalType,
  fromX,
  fromY,
  quantizationWidth = NA,
  verbose = FALSE
)
}
\arguments{
\item{map}{A PointMap, Axial ShapeGraph or Segment ShapeGraph}

\item{traversalType}{The traversal type. See \link{TraversalType}}

\item{fromX}{X coordinate of the point to start the traversal from}

\item{fromY}{X coordinate of the point to start the traversal from}

\item{quantizationWidth}{Set this to use chunks of this width instead of
continuous values for the cost of traversal. This is equivalent to the "tulip
bins" for depthmapX's tulip analysis (1024 tulip bins = pi/1024
quantizationWidth). Only works for Segment ShapeGraphs}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
Returns a list with:
\itemize{
  \item{completed: Whether the analysis completed}
  \item{newAttributes: The new attributes that were created during the
  process}
}
}
\description{
Runs one-to-all traversal on a map with a graph. This is applicable to:
\itemize{
  \item{PointMaps (Visibility Graph Analysis)}
  \item{Axial ShapeGraphs (Axial analysis)}
  \item{Segment ShapeGraphs (Segment analysis)}
}
}
