% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palettes}
\alias{palettes}
\alias{depthmap.classic.colour}
\alias{depthmap.axmanesque.colour}
\alias{depthmap.purpleorange.colour}
\alias{depthmap.bluered.colour}
\alias{depthmap.grayscale.colour}
\alias{depthmap.nicehsb.colour}
\title{Colour Palettes from depthmapX}
\usage{
depthmap.classic.colour(n, rangeMin = 0, rangeMax = 1)

depthmap.axmanesque.colour(n, rangeMin = 0, rangeMax = 1)

depthmap.purpleorange.colour(n, rangeMin = 0, rangeMax = 1)

depthmap.bluered.colour(n, rangeMin = 0, rangeMax = 1)

depthmap.grayscale.colour(n, rangeMin = 0, rangeMax = 1)

depthmap.nicehsb.colour(n, rangeMin = 0, rangeMax = 1)
}
\arguments{
\item{n}{Number of colours to generate}

\item{rangeMin}{The min value of the range}

\item{rangeMax}{The max value of the range}
}
\value{
Returns a vector of colours.
}
\description{
Create n contiguous colours taken from depthmapX.
}
\examples{
depthmap.classic.colour(100, 0, 1)
depthmap.axmanesque.colour(100, 0, 1)
depthmap.purpleorange.colour(100, 0, 1)
depthmap.bluered.colour(100, 0, 1)
depthmap.grayscale.colour(100, 0, 1)
depthmap.nicehsb.colour(100, 0, 1)
}
