% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeMap.R
\name{plot.LatticeMap}
\alias{plot.LatticeMap}
\title{plot a LatticeMap}
\usage{
\method{plot}{LatticeMap}(x, ...)
}
\arguments{
\item{x}{object of class \code{LatticeMap}}

\item{...}{other parameters passed to \code{stars[]}}
}
\description{
Calls a standard plot.stars, but flips the first argument around the x axis
}
