% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\name{plot.PointMap}
\alias{plot.PointMap}
\title{plot a PointMap}
\usage{
\method{plot}{PointMap}(x, ...)
}
\arguments{
\item{x}{object of class \code{PointMap}}

\item{...}{other parameters passed to \code{stars[]}}
}
\description{
Calls a standard plot.stars, but flips the first argument around the x axis
}
