% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\name{PointMap-class}
\alias{PointMap-class}
\alias{PointMap}
\title{PointMap}
\description{
A representation of sala's PointMap in R. Holds onto a sala PointMap pointer
and operates on that
}
\concept{PointMap}
