% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\name{PointMap_subset}
\alias{PointMap_subset}
\alias{[.PointMap}
\alias{[<-.PointMap}
\title{Subset PointMap objects}
\usage{
\method{[}{PointMap}(x, ...)

\method{[}{PointMap}(x, ...) <- value
}
\arguments{
\item{x}{object of class \code{PointMap} passed to \code{stars[]}}

\item{...}{other parameters passed to \code{stars[] <- }}

\item{value}{value to be passed to \code{stars[] <- }}
}
\description{
Subsetting PointMap objects essentially passes the data to stars
See \link[stars]{stars_subset}
}
