% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMetaGraph.R
\name{readMetaGraph}
\alias{readMetaGraph}
\title{Read MetaGraph}
\usage{
readMetaGraph(fileName)
}
\arguments{
\item{fileName}{The metagraph file}
}
\value{
A list of ShapeMaps, ShapeGraphs and PointMaps
}
\description{
Reads a metagraph into a bunch of ShapeMaps/ShapeGraphs/PointMaps
}
\examples{
fileName <- system.file(
  "extdata", "testdata", "barnsbury", "barnsburySmall.graph",
  package = "alcyon"
)
readMetaGraph(fileName)
}
