% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refIDtoIndex.R
\name{refIdToIndexAndBack}
\alias{refIdToIndexAndBack}
\alias{refIDtoIndex}
\alias{indexToRefID}
\title{Ref ID to index and vice-versa}
\usage{
refIDtoIndex(refID)

indexToRefID(i, j)
}
\arguments{
\item{refID}{The Ref ID}

\item{i}{The x-axis index of the cell}

\item{j}{The y-axis index of the cell}
}
\value{
A pair of indices (x, y) or a Ref ID
}
\description{
Converts a depthmapX "Ref" ID to the indices (x, y) of a cell, or the reverse
}
\examples{
idx <- refIDtoIndex(852645)
# outputs:
#    i   j
# 1 13 677

idx <- indexToRefID(13, 667)
# outputs:
# 852645
}
