% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentShapeGraph.R
\name{SegmentShapeGraph_subset}
\alias{SegmentShapeGraph_subset}
\alias{[.SegmentShapeGraph}
\alias{[<-.SegmentShapeGraph}
\title{Subset SegmentShapeGraph objects}
\usage{
\method{[}{SegmentShapeGraph}(x, ...)

\method{[}{SegmentShapeGraph}(x, ...) <- value
}
\arguments{
\item{x}{object of class \code{SegmentShapeGraph} passed to \code{stars[]}}

\item{...}{other parameters passed to \code{stars[] <- }}

\item{value}{value to be passed to \code{sf[] <- }}
}
\description{
Subsetting SegmentShapeGraph objects essentially passes the data to sf.
See \link[sf]{sf}
}
