% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeMap.R
\name{ShapeMap_subset}
\alias{ShapeMap_subset}
\alias{[.ShapeMap}
\alias{[<-.ShapeMap}
\title{Subset ShapeMap objects}
\usage{
\method{[}{ShapeMap}(x, ...)

\method{[}{ShapeMap}(x, ...) <- value
}
\arguments{
\item{x}{object of class \code{ShapeMap} passed to \code{sf[]}}

\item{...}{other parameters passed to \code{sf[] <- }}

\item{value}{value to be passed to \code{sf[] <- }}
}
\description{
Subsetting ShapeMap objects essentially passes the data to sf.
See \link[sf]{sf}
}
