% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfConversions.R
\name{shapeMapToPolygonSf}
\alias{shapeMapToPolygonSf}
\title{ShapeMap to sf Polygon map}
\usage{
shapeMapToPolygonSf(shapeMap)
}
\arguments{
\item{shapeMap}{A ShapeMap}
}
\value{
A new sf Polygon map
}
\description{
Convert a ShapeMap to an sf Polygon map
}
