% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraversalType.R
\name{TraversalType}
\alias{TraversalType}
\title{Traversal types}
\usage{
TraversalType
}
\value{
A list of numbers representing each particular analysis type
}
\description{
These are meant to be used to indicate what kind of analysis
is meant to be carried out.
}
\examples{
TraversalType$Angular
TraversalType$Topological
TraversalType$Metric
}
