% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{unlinkAtCrossPoint}
\alias{unlinkAtCrossPoint}
\title{Unlink map lines at their crossing point}
\usage{
unlinkAtCrossPoint(map, x, y)
}
\arguments{
\item{map}{A sala map}

\item{x}{X coordinate of the crossing point}

\item{y}{Y coordinate of the crossing point}
}
\value{
None
}
\description{
Unlink map lines at their crossing point
}
