% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxialShapeGraph.R
\docType{methods}
\name{unlinkCoords,AxialShapeGraph-method}
\alias{unlinkCoords,AxialShapeGraph-method}
\title{Unlink two Axial Lines (coordinates)}
\usage{
\S4method{unlinkCoords}{AxialShapeGraph}(map, fromX, fromY, toX, toY)
}
\arguments{
\item{map}{An Axial ShapeGraph}

\item{fromX}{X coordinate of the first unlink point}

\item{fromY}{Y coordinate of the first unlink point}

\item{toX}{X coordinate of the second unlink point}

\item{toY}{Y coordinate of the second unlink point}
}
\description{
Unlink two locations on an Axial ShapeGraph using the point coordinates
}
