% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\docType{methods}
\name{unlinkCoords,PointMap-method}
\alias{unlinkCoords,PointMap-method}
\title{Unlink two PointMap Cells (coordinates)}
\usage{
\S4method{unlinkCoords}{PointMap}(map, fromX, fromY, toX, toY)
}
\arguments{
\item{map}{A PointMap}

\item{fromX}{X coordinate of the first unlink point}

\item{fromY}{Y coordinate of the first unlink point}

\item{toX}{X coordinate of the second unlink point}

\item{toY}{Y coordinate of the second unlink point}
}
\description{
Unlink two cells on a PointMap using the point coordinates
}
