% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatticeMap.R
\docType{methods}
\name{unlinkRefs,LatticeMap-method}
\alias{unlinkRefs,LatticeMap-method}
\title{Unlink two LatticeMap Cells (refs)}
\usage{
\S4method{unlinkRefs}{LatticeMap}(map, fromRef, toRef, copyMap = TRUE)
}
\arguments{
\item{map}{A LatticeMap}

\item{fromRef}{Ref of the first unlink line}

\item{toRef}{Ref of the second unlink line}

\item{copyMap}{Optional. Copy the internal sala map}
}
\value{
A new LatticeMap with unlinked points
}
\description{
Unlink two cells on an LatticeMap using their refs
}
\examples{
mifFile <- system.file(
    "extdata", "testdata", "gallery",
    "gallery_lines.mif",
    package = "alcyon"
  )
  sfMap <- st_read(mifFile,
    geometry_column = 1L, quiet = TRUE
  )
  latticeMap <- makeVGALatticeMap(
    sfMap,
    gridSize = 0.04,
    fillX = 3.01,
    fillY = 6.7,
    maxVisibility = NA,
    boundaryGraph = FALSE,
    verbose = FALSE
  )
latticeMap <- linkRefs(latticeMap, 1835056L, 7208971L)
latticeMap <- unlinkRefs(latticeMap, 1835056L, 7208971L)
}
