% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\docType{methods}
\name{unlinkRefs,PointMap-method}
\alias{unlinkRefs,PointMap-method}
\title{Unlink two Axial Lines (refs)}
\usage{
\S4method{unlinkRefs}{PointMap}(map, fromRef, toRef)
}
\arguments{
\item{map}{A PointMap}

\item{fromRef}{Ref of the first unlink line}

\item{toRef}{Ref of the second unlink line}
}
\description{
Unlink two lines on an Axial ShapeGraph using their refs
}
