% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{unlinkRefs}
\alias{unlinkRefs}
\title{Unlink map points/lines using their refs}
\usage{
unlinkRefs(map, fromRef, toRef, copyMap = TRUE)
}
\arguments{
\item{map}{A sala map}

\item{fromRef}{The ref of the origin element}

\item{toRef}{The ref of the target element}

\item{copyMap}{Optional. Copy the internal sala map}
}
\value{
A new map with unlinked points/lines
}
\description{
Unlink map points/lines using their refs
}
