% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVGA.R
\name{unmakeVGAGraph}
\alias{unmakeVGAGraph}
\title{Unmake the graph in a PointMap}
\usage{
unmakeVGAGraph(pointMap, removeLinks = FALSE, verbose = FALSE)
}
\arguments{
\item{pointMap}{The input PointMap}

\item{removeLinks}{Also remove the links}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
A new PointMap
}
\description{
Unmake the graph in a PointMap
}
