% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allToAllTraverse.R
\name{vgaIsovist}
\alias{vgaIsovist}
\title{Visibility Graph Analysis - isovist metrics}
\usage{
vgaIsovist(pointMap, boundaryMap)
}
\arguments{
\item{pointMap}{A PointMap}

\item{boundaryMap}{A ShapeMap of lines}
}
\value{
None
}
\description{
Runs axial analysis to get the local metrics Control and Controllability
}
\examples{
mifFile = system.file(
    "extdata", "testdata", "gallery",
    "gallery_lines.mif",
    package = "alcyon"
  )
  sfMap <- st_read(mifFile,
    geometry_column = 1L, quiet = TRUE
  )
  pointMap <- makeVGAPointMap(
    sfMap,
    gridSize = 0.04,
    fillX = 3.01,
    fillY = 6.7,
    maxVisibility = NA,
    boundaryGraph = FALSE,
    verbose = FALSE
  )
boundaryMap <- as(sfMap[, c()], "ShapeMap")
vgaIsovist(pointMap, boundaryMap)
}
