% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allToAllTraverse.R
\name{vgaThroughVision}
\alias{vgaThroughVision}
\title{Visibility Graph Analysis - Through Vision}
\usage{
vgaThroughVision(pointMap)
}
\arguments{
\item{pointMap}{A PointMap}
}
\description{
Runs Visibility Graph Analysis to get the Through Vision metric
}
