% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allToAllTraverse.R
\name{vgaVisualLocal}
\alias{vgaVisualLocal}
\title{Visibility Graph Analysis - Visual local metrics}
\usage{
vgaVisualLocal(pointMap, gatesOnly = FALSE)
}
\arguments{
\item{pointMap}{A PointMap}

\item{gatesOnly}{Optional. Only keep the values at specific gates}
}
\description{
Runs Visibility Graph Analysis to get visual local metrics
}
