% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentAnalysis.R
\name{agentAnalysis}
\alias{agentAnalysis}
\title{Agent Analysis}
\usage{
agentAnalysis(
  pointMap,
  timesteps,
  releaseRate,
  agentLifeTimesteps,
  agentFov,
  agentStepsToDecision,
  agentLookMode,
  originX = NA,
  originY = NA,
  locationSeed = 0L,
  numberOfTrails = NA,
  getGateCounts = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{pointMap}{A PointMap, used as an exosomatic visual map for agents to
take exploratory information}

\item{timesteps}{Number of total system timesteps.}

\item{releaseRate}{Agent release rate (likelyhood of release per timestep).}

\item{agentLifeTimesteps}{Agent total lifetime (in timesteps)}

\item{agentFov}{Agent field-of-view (out of 32 bins = 360).}

\item{agentStepsToDecision}{Agent steps before turn decision.}

\item{agentLookMode}{The agent look mode. See \link{AgentLookMode}}

\item{originX}{Agent starting points (x coordinates).}

\item{originY}{Agent starting point (y coordinates).}

\item{locationSeed}{Agents to start at random locations with specific seed (0
to 10). Default is 0.}

\item{numberOfTrails}{Record trails for this amount of agents (set to 0 to
record all, with max possible currently = 50).}

\item{getGateCounts}{Get values at gates}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
Returns a list with:
\itemize{
  \item{completed: Whether the analysis completed}
  \item{newAttributes: The new attributes that were created during the
  process}
}
}
\description{
Runs Agent Analysis on the given PointMap
}
