% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeMap.R, R/AxialShapeGraph.R,
%   R/SegmentShapeGraph.R
\name{as}
\alias{as}
\title{as("sf", "ShapeMap")}
\description{
This is a direct conversion, for ShapeMap -> Axial -> Segment see
\link{axialToSegmentShapeGraph}

This is a direct conversion, for ShapeMap -> Axial -> Segment see
\link{axialToSegmentShapeGraph}
}
\seealso{
Other ShapeMap: 
\code{\link{ShapeMap-class}}

Other ShapeMap: 
\code{\link{ShapeMap-class}}

Other AxialShapeGraph: 
\code{\link{AxialShapeGraph-class}}

Other AxialShapeGraph: 
\code{\link{AxialShapeGraph-class}}

Other SegmentShapeGraph: 
\code{\link{SegmentShapeGraph-class}}

Other SegmentShapeGraph: 
\code{\link{SegmentShapeGraph-class}}
}
\concept{AxialShapeGraph}
\concept{SegmentShapeGraph}
\concept{ShapeMap}
