% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeMap.R, R/AxialShapeGraph.R,
%   R/SegmentShapeGraph.R, R/PointMap.R
\name{as}
\alias{as}
\title{as("sf", "ShapeMap")}
\description{
This is a direct conversion, for ShapeMap -> Axial -> Segment see
\link{axialToSegmentShapeGraph}
}
\concept{AxialShapeGraph}
\concept{PointMap}
\concept{SegmentShapeGraph}
\concept{ShapeMap}
