% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxialShapeGraph.R
\docType{class}
\name{AxialShapeGraph-class}
\alias{AxialShapeGraph-class}
\title{Axial ShapeGraph}
\description{
A representation of sala's Axial ShapeGraph in R. Holds onto a sala Axial
ShapeGraph pointer and operates on that
}
