% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVGA.R
\name{createGrid}
\alias{createGrid}
\title{Create an emprt PointMap through a grid}
\usage{
createGrid(minX, minY, maxX, maxY, gridSize, verbose = FALSE)
}
\arguments{
\item{minX}{Minimum X of the bounding region}

\item{minY}{Minimum Y of the bounding region}

\item{maxX}{Maximum X of the bounding region}

\item{maxY}{Maximum Y of the bounding region}

\item{gridSize}{Size of the cells}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
A new PointMap
}
\description{
Create an emprt PointMap through a grid
}
