% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVGA.R
\name{fillGrid}
\alias{fillGrid}
\title{Fill a PointMap's grid starting from one or more points}
\usage{
fillGrid(pointMap, fillX, fillY, verbose = FALSE)
}
\arguments{
\item{pointMap}{The input PointMap}

\item{fillX}{X coordinate of the fill points}

\item{fillY}{Y coordinate of the fill points}

\item{verbose}{Optional. Show more information of the process.}
}
\description{
Fill a PointMap's grid starting from one or more points
}
