% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointMap.R
\docType{methods}
\name{linkRefs,PointMap-method}
\alias{linkRefs,PointMap-method}
\title{Link two PointMap Cells (refs)}
\usage{
\S4method{linkRefs}{PointMap}(map, fromRef, toRef, copyMap = TRUE)
}
\arguments{
\item{map}{A PointMap}

\item{fromRef}{Ref of the first link line}

\item{toRef}{Ref of the second link line}

\item{copyMap}{Optional. Copy the internal sala map}
}
\value{
A new PointMap with linked points
}
\description{
Link two cells on an PointMap using their refs
}
\examples{
mifFile <- system.file(
    "extdata", "testdata", "gallery",
    "gallery_lines.mif",
    package = "alcyon"
  )
  sfMap <- st_read(mifFile,
    geometry_column = 1L, quiet = TRUE
  )
  pointMap <- makeVGAPointMap(
    sfMap,
    gridSize = 0.04,
    fillX = 3.01,
    fillY = 6.7,
    maxVisibility = NA,
    boundaryGraph = FALSE,
    verbose = FALSE
  )
pointMap <- linkRefs(pointMap, 1835056L, 7208971L)
}
