% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVGA.R
\name{makeVGAGraph}
\alias{makeVGAGraph}
\title{Create a graph between visible cells in the LatticeMap}
\usage{
makeVGAGraph(
  latticeMap,
  boundaryGraph = FALSE,
  maxVisibility = NA,
  copyMap = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{latticeMap}{The input LatticeMap}

\item{boundaryGraph}{Only create a graph on the boundary cells}

\item{maxVisibility}{Limit how far two cells can be to be connected}

\item{copyMap}{Optional. Copy the internal sala map}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
A new LatticeMap with a graph between points
}
\description{
Create a graph between visible cells in the LatticeMap
}
\examples{
mifFile <- system.file(
    "extdata", "testdata", "simple",
    "simple_interior.mif",
    package = "alcyon"
  )
  sfMap <- st_read(mifFile,
    geometry_column = 1L, quiet = TRUE
  )
  shapeMap <- as(sfMap[, vector()], "ShapeMap")
lineStringMap <- as(sfMap, "sf")
mapRegion <- sf::st_bbox(lineStringMap)
latticeMap <- createGrid(
  minX = mapRegion[["xmin"]],
  minY = mapRegion[["ymin"]],
  maxX = mapRegion[["xmax"]],
  maxY = mapRegion[["ymax"]],
  gridSize = 0.5
)
latticeMap <- blockLines(
  latticeMap = latticeMap,
  lineStringMap = lineStringMap[vector()]
)
latticeMap <- fillGrid(
  latticeMap = latticeMap,
  fillX = 3.01,
  fillY = 6.7
)
makeVGAGraph(
  latticeMap = latticeMap,
  boundaryGraph = FALSE,
  maxVisibility = NA
)
}
