% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allFewestLineMap.R
\name{reduceToFewest}
\alias{reduceToFewest}
\title{Reduce an All-line Map to two types of fewest-line maps}
\usage{
reduceToFewest(allLineMap)
}
\arguments{
\item{allLineMap}{An AllLineShapeGraph}
}
\value{
A list with two fewest-line axial ShapeGraphs
}
\description{
Reduce an All-line Map to two types of fewest-line maps
}
