% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentShapeGraph.R
\name{SegmentShapeGraph-class}
\alias{SegmentShapeGraph-class}
\alias{SegmentShapeGraph}
\title{Segment ShapeGraph}
\description{
A representation of sala's Segment ShapeGraph in R. Holds onto a sala Segment
ShapeGraph pointer and operates on that
}
\seealso{
Other SegmentShapeGraph: 
\code{\link{as}()}
}
\concept{SegmentShapeGraph}
