% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapegraphToIGraph.R
\name{shapegraphToIGraph}
\alias{shapegraphToIGraph}
\title{Conversion of shapegraph to igraph}
\usage{
shapegraphToIGraph(shapeGraph, weightColumn = NA)
}
\arguments{
\item{shapeGraph}{A ShapeGraph}

\item{weightColumn}{Optional.The variable used to assign weight to graph
edges}
}
\value{
Returns graph.data.frame.
}
\description{
Creates igraph based on the connections and the x,y coordinates of the
centroids of shapes in a shapegraph (axial,segment, convex). Specify
weightColumn to assign weight to graph edges.
}
\details{
If weightColumn is provided, edge connections weight is calculated by taking
the average of the variable of the connected nodes.
}
