% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeMap.R
\name{ShapeMap-class}
\alias{ShapeMap-class}
\alias{ShapeMap}
\title{ShapeMap class}
\description{
A representation of sala's ShapeMap in R. Holds onto a sala ShapeMap
pointer and operates on that
}
\seealso{
Other ShapeMap: 
\code{\link{as}()}
}
\concept{ShapeMap}
