% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{unlinkRefs}
\alias{unlinkRefs}
\title{Unlink map cells/lines using their refs}
\usage{
unlinkRefs(map, fromRef, toRef)
}
\arguments{
\item{map}{A sala map}

\item{fromRef}{The ref of the origin element}

\item{toRef}{The ref of the target element}
}
\value{
None
}
\description{
Unlink map cells/lines using their refs
}
