% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareVGA.R
\name{unmakeVGAGraph}
\alias{unmakeVGAGraph}
\title{Unmake the graph in a LatticeMap}
\usage{
unmakeVGAGraph(
  latticeMap,
  removeLinks = FALSE,
  copyMap = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{latticeMap}{The input LatticeMap}

\item{removeLinks}{Also remove the links}

\item{copyMap}{Optional. Copy the internal sala map}

\item{verbose}{Optional. Show more information of the process.}
}
\value{
A new LatticeMap without the points graph
}
\description{
Unmake the graph in a LatticeMap
}
\examples{
mifFile <- system.file(
    "extdata", "testdata", "simple",
    "simple_interior.mif",
    package = "alcyon"
  )
  sfMap <- st_read(mifFile,
    geometry_column = 1L, quiet = TRUE
  )
  shapeMap <- as(sfMap[, vector()], "ShapeMap")
latticeMap <- makeVGALatticeMap(
  sfMap,
  gridSize = 0.5,
  fillX = 3.01,
  fillY = 6.7,
  maxVisibility = NA,
  boundaryGraph = FALSE,
  verbose = FALSE
)
unmakeVGAGraph(
  latticeMap = latticeMap,
  removeLinks = FALSE
)
}
