% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allToAllTraverse.R
\name{vgaIsovist}
\alias{vgaIsovist}
\title{Visibility Graph Analysis - isovist metrics}
\usage{
vgaIsovist(pointMap, boundaryMap)
}
\arguments{
\item{pointMap}{A PointMap}

\item{boundaryMap}{A ShapeMap of lines}
}
\description{
Runs axial analysis to get the local metrics Control and Controllability
}
