% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgaLocal.R
\docType{data}
\name{VGALocalAlgorithm}
\alias{VGALocalAlgorithm}
\title{VGA Local Analysis algorithms.}
\format{
An object of class \code{list} of length 3.
}
\usage{
VGALocalAlgorithm
}
\value{
A list of numbers representing each algorithm
}
\description{
Different algorithms for calculating the VGA Local metrics (Control,
Controllability, Clustering Coefficient).
\itemize{
  \item{VGALocalAlgorithm$None}
  \item{VGALocalAlgorithm$Standard}
  \item{VGALocalAlgorithm$AdjacencyMatrix}
}
}
\examples{
VGALocalAlgorithm$Angular
VGALocalAlgorithm$Topological
VGALocalAlgorithm$Metric
}
\keyword{datasets}
